requirements.txt
anthropic
python-dotenv
pydantic
fastapi
uvicorn
supabase
httpx
slowapi
replicate
